+++ GoodBuyUS/src/main/java/dev_final_team10/GoodBuyUS/controller/CommunityPaymentController.java
@@ -4,6 +4,7 @@ import dev_final_team10.GoodBuyUS.domain.payment.dto.CommunityPaymentRequestDto;
 import dev_final_team10.GoodBuyUS.domain.payment.dto.CommunityPaymentResponseDto;
 import dev_final_team10.GoodBuyUS.domain.payment.dto.TossWebhookDto;
 import dev_final_team10.GoodBuyUS.service.CommunityPaymentService;
+import dev_final_team10.GoodBuyUS.service.CommunityWebhookService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -15,9 +16,12 @@ import java.util.Map;
 public class CommunityPaymentController {
 
     private final CommunityPaymentService communityPaymentService;
+    private final CommunityWebhookService communityWebhookService;
 
-    public CommunityPaymentController(CommunityPaymentService communityPaymentService) {
+    public CommunityPaymentController(CommunityPaymentService communityPaymentService,
+                                        CommunityWebhookService communityWebhookService) {
         this.communityPaymentService = communityPaymentService;
+        this.communityWebhookService = communityWebhookService;
     }
 
     @PostMapping
@@ -107,17 +111,17 @@ public class CommunityPaymentController {
                     .body(Map.of("error", "결제 상태 조회 중 오류 발생", "message", e.getMessage()));
         }
     }
-/* 웹훅 기능 사용시 커뮤니티 결제 서비스에서 웹훅 주석 푸시고 이것도 푸시면 됩니다.
+ //웹훅 기능 사용시 커뮤니티 결제 서비스에서 웹훅 주석 푸시고 이것도 푸시면 됩니다.
     @PostMapping("/webhook")
     public ResponseEntity<?> handleWebhook(@RequestBody TossWebhookDto webhookDto) {
         try {
-            communityPaymentService.processWebhook(webhookDto);
+            communityWebhookService.processWebhook(webhookDto);
             return ResponseEntity.ok("Webhook processed successfully");
         } catch (Exception e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
                     .body(Map.of("error", e.getMessage()));
         }
-    }*/
+    }
 }
