+++ GoodBuyUS/src/main/resources/static/index.html
@@ -1,0 +1,24 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>SSE Test</title>
+</head>
+<body>
+<h1>SSE 테스트</h1>
+<script>
+    const paymentKey = "test_payment_key";
+    const eventSource = new EventSource("http://localhost:8080/api/v1/sse/events/" + paymentKey);
+
+    eventSource.onmessage = (event) => {
+        console.log("SSE 메시지:", event.data);
+        alert("SSE 메시지: " + event.data);
+    };
+
+    eventSource.onerror = (error) => {
+        console.error("SSE 오류 발생:", error);
+    };
+</script>
+</body>
+</html>
